Update guest login
